Split remote service apis into separate files.
